refactor: define SessionLookup trait in closeclaw-common
Extract PendingMessage from closeclaw-session into closeclaw-common
and define the SessionLookup async trait with get_parent_of,
get_chat_id, and push_pending_message methods. This decouples the
permission module from the gateway's SessionManager.

Source: issue #1252
Type: refactor
