+++ crates/common/Cargo.toml
@@ -6,6 +6,8 @@ description = "Shared types and utilities for CloseClaw"
 license = "MIT"
 
 [dependencies]
+async-trait = "0.1"
+chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "2.0"
