+++ crates/common/src/lib.rs
@@ -2,6 +2,7 @@ pub mod agent_config;
 pub mod bootstrap;
 pub mod communication;
 pub mod compaction;
+pub mod session_lookup;
 pub mod session_types;
 pub mod shutdown;
 pub mod verbosity;
@@ -15,6 +16,7 @@ pub use communication::{
     check_communication_allowed, CommunicationCheckResult, CommunicationConfig, CommunicationError,
 };
 pub use compaction::CompactConfig;
+pub use session_lookup::{PendingMessage, SessionLookup};
 pub use session_types::{AgentRole, ReasoningLevel};
 pub use shutdown::ShutdownSignal;
 pub use verbosity::VerbosityLevel;
